fixing test failures
- handle "-sync" option in QuickLauncher and make the default as false
- skip waiting for ExternalCatalog in smart connector mode (where it will be absent)
- set back password, if set, when initializing the global hive catalog
- deal with removal of embedded lead creation using SparkContext in ClusterMgrDUnitTest
- add "jobserver.waitForInitialization=true" for hydra tests

+++ cluster/src/dunit/scala/io/snappydata/cluster/ClusterManagerTestBase.scala
@@ -293,6 +293,7 @@ object ClusterManagerTestBase extends Logging {
 
   def cleanupTestData(testClass: String, testName: String): Unit = {
     // cleanup metastore
+    if (Misc.getMemStoreBootingNoThrow eq null) return
     val snc = SnappyContext()
     if (snc != null) {
       TestUtils.dropAllTables(snc)
